TWID-SOA/control/requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
pydantic==1.9.1
python-dotenv==0.21.0
requests==2.28.1